+++ extension/src/cli/git/constants.ts
@@ -1,3 +1,12 @@
+import { join } from 'path'
+
+export const DOT_GIT = '.git'
+export const DOT_GIT_HEAD = join(DOT_GIT, 'HEAD')
+export const DOT_GIT_INDEX = join(DOT_GIT, 'index')
+export const GIT_REFS = join(DOT_GIT, 'refs')
+export const GIT_LOGS_REFS = join(DOT_GIT, 'logs', 'refs')
+export const HEADS_GIT_REFS = join(GIT_REFS, 'heads')
+
 export enum Command {
   ADD = 'add',
   CLEAN = 'clean',
