+++ extension/src/experiments/data/constants.ts
@@ -1,5 +1,5 @@
 import { join } from 'path'
-import { GIT_LOGS_REFS, GIT_REFS } from '../../git'
+import { GIT_LOGS_REFS, GIT_REFS } from '../../cli/git/constants'
 
 export const EXPERIMENTS_GIT_REFS = join(GIT_REFS, 'exps')
 export const EXPERIMENTS_GIT_LOGS_REFS = join(GIT_LOGS_REFS, 'exps')
